# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/dashboard/summary": {
      "get": {
        "operationId": "DashboardDataController_summary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DashboardSummaryDto"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Dashboard data is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Counts for the dashboard summary card",
        "tags": [
          "dashboard-data"
        ]
      }
    },
...
    "/api/dashboards/{dashboardId}": {
      "get": {
        "description": "A dashboard the caller may not see answers 404, not 403 — a distinct 403 would confirm the id exists.",
        "operationId": "DashboardController_findOne",
        "parameters": [
          {
            "name": "dashboardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DashboardResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such dashboard, or not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read a dashboard definition",
        "tags": [
          "dashboards"
        ]
      },
      "put": {
        "description": "Requires the version last read. A mismatch answers 409 with the current version in details.",
        "operationId": "DashboardController_save",
        "parameters": [
          {
            "name": "dashboardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDashboardDefinitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DashboardResponseDto"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The caller can see the dashboard but does not own it.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "The dashboard changed since it was read.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "The definition failed schema validation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace a dashboard definition",
        "tags": [
          "dashboards"
        ]
      }
    },
    "/api/dashboards/{dashboardId}/personalization": {
      "get": {
        "description": "Creates the default personalization on first access rather than answering 404.",
        "operationId": "DashboardController_findPersonalization",
        "parameters": [
          {
            "name": "dashboardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PersonalizationResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such dashboard, or not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read the caller's personalization",
        "tags": [
          "dashboards"
        ]
      },
      "put": {
        "operationId": "DashboardController_savePersonalization",
        "parameters": [
          {
            "name": "dashboardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePersonalizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PersonalizationResponseDto"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "The personalization changed since it was read.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "The presets failed schema validation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace the whole personalization",
        "tags": [
          "dashboards"
        ]
      }
    },
    "/api/dashboards/{dashboardId}/presets": {
      "post": {
        "description": "Copies the active preset unless another is named.",
        "operationId": "DashboardController_createPreset",
        "parameters": [
          {
            "name": "dashboardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePresetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PersonalizationResponseDto"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "The personalization changed since it was read.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a preset",
        "tags": [
          "dashboards"
        ]
      }
    },
...
    "/api/dashboards/{dashboardId}/presets/{presetId}/select": {
      "post": {
        "operationId": "DashboardController_selectPreset",
        "parameters": [
          {
            "name": "dashboardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "presetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectPresetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PersonalizationResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such preset.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Make a preset the active one",
        "tags": [
          "dashboards"
        ]
      }
    },
...
  },
  "info": {
    "title": "React Boilerplate API",
    "description": "Backend for the React boilerplate. Every response uses the shared envelope: `{ success: true, data }` on success and `{ success: false, error: { code, message, details? } }` on failure. HTTP status describes the transport outcome; `error.code` describes the domain outcome.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "rb_refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rb_refresh"
      }
    },
    "schemas": {
...
      "DashboardSummaryDto": {
        "type": "object",
        "properties": {
          "activeUsers": {
            "type": "number"
          },
          "invitedUsers": {
            "type": "number"
          },
          "blockedUsers": {
            "type": "number"
          },
          "apiLatencyMs": {
            "type": "number"
          },
          "contractErrorRate": {
            "type": "number"
          }
        },
        "required": [
          "activeUsers",
          "invitedUsers",
          "blockedUsers",
          "apiLatencyMs",
          "contractErrorRate"
        ]
      },
...
      "DashboardResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "shared"
            ]
          },
          "schemaVersion": {
            "type": "number",
            "description": "Schema version of the stored definition."
          },
          "version": {
            "type": "number",
            "description": "Optimistic lock. Send this back when writing."
          },
          "definition": {
            "type": "object",
            "description": "Validated dashboard definition."
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "ownerId",
          "visibility",
          "schemaVersion",
          "version",
          "definition",
          "updatedAt"
        ]
      },
      "SaveDashboardDefinitionDto": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "number"
          },
          "definition": {
            "type": "object"
          }
        },
        "required": [
          "expectedVersion",
          "definition"
        ]
      },
...
      "SavePersonalizationDto": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "number",
            "description": "The version last read. A mismatch answers 409."
          },
          "activePresetId": {
            "type": "string"
          },
          "presets": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "expectedVersion",
          "activePresetId",
          "presets"
        ]
      },
...
    }
  }
}